Workaround for HIBP check in Home Assistant
As of Dec 30 2025, Home Assistant checks passwords for leaks on https://haveibeenpwned.com/ and if this check fails for some reason (connection issues, HIBP outage, government restrictions in some countries, etc.), it refuses to save the config or start the addon. Even though Home Assistant devs provided a way to disable this check, it doesn't disable it completely.